# services/gateway-api/openapi.yaml
openapi: "3.1.0"
info:
  title: Opencomplai Gateway API
  version: 0.9.0
  description: |
    REST API gateway for the Opencomplai compliance platform.
    All routes proxy to the appropriate Python service.
    This spec is the source of truth for the API contract.

    Manifests can name several frameworks (compliance_targets). Risk
    classification, dossier generation and the risk engine's public
    /v1/checker/* routes (called by the docs widget, not proxied here) are
    EU AI Act only.

    Authentication: every /v1 route requires credentials, either an
    x-api-key header (API-key mode) or an Authorization bearer JWT (OIDC
    mode). A deployment uses one of the two. Only GET /health is open. The
    gateway refuses to start unconfigured unless authentication is explicitly
    disabled for local development (OPENCOMPLAI_AUTH_DISABLED=1).
servers:
  - url: http://localhost:8080
    description: Local Docker Compose deployment
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /v1/evidence/events:
    post:
      operationId: appendEvidenceEvent
      summary: Append an event to the immutable ledger (non-idempotent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppendEventRequest"
      responses:
        "201":
          description: Event appended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppendEventResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /v1/evidence/ledger-root:
    get:
      operationId: getLedgerRoot
      summary: Current Merkle chain tip of the tenant's evidence ledger (idempotent)
      description: >-
        The hash an Annex IV dossier anchors to. For an empty ledger this is
        the genesis hash (sha256 of the empty string).
      responses:
        "200":
          description: Current chain tip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerRootResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
...
  /v1/docs/generate:
    post:
      operationId: generateDocs
      summary: Generate Annex IV technical documentation dossier (EU AI Act only; idempotent by system_id + commit_ref)
      description: >-
        Builds the EU AI Act Annex IV dossier; other frameworks have no
        dossier. The gateway forwards the body unchanged to the doc generator,
        which validates it. The dossier is stored in the evidence vault and
        can be fetched with GET /v1/docs/{dossier_id}. By default a dossier
        that fails the Annex IV schema check is still stored but the request
        is refused with 422 DOSSIER_SCHEMA_INVALID; set allow_incomplete to
        accept it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDocsRequest"
      responses:
        "200":
          description: Dossier generated and stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateDocsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: >-
            The dossier failed schema validation and allow_incomplete was not
            set (DOSSIER_SCHEMA_INVALID; the dossier is stored and its ids
            are in the body), eval_report or corroboration_report is not a
            valid report (INVALID_EVAL_REPORT, INVALID_CORROBORATION_REPORT),
            or the body failed request validation (detail is a list)
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/DossierSchemaInvalidError"
                  - $ref: "#/components/schemas/UpstreamError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          description: >-
            Generation failed unexpectedly (SYSTEM_ERROR, with the message and
            duration_ms), for example an unknown compliance_target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "502":
          description: >-
            The dossier was generated but could not be stored in the evidence
            vault (VAULT_CAS_WRITE_FAILED, VAULT_EVENT_APPEND_FAILED or
            VAULT_DOSSIER_INDEX_FAILED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /v1/docs:
    get:
      operationId: listDossiers
      summary: List the dossiers stored for a system, newest first (idempotent)
      parameters:
        - name: system_id
          in: query
          required: true
          schema:
            type: string
          description: System whose dossiers to list
      responses:
        "200":
          description: Dossier index entries for the system
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DossierListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: system_id query parameter missing or empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          description: The dossier index could not be read from the evidence vault
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
  /v1/docs/{dossier_id}:
    get:
      operationId: getDossier
      summary: Retrieve a generated dossier and its index metadata (idempotent)
      parameters:
        - name: dossier_id
          in: path
          required: true
          schema:
            type: string
          description: Dossier id returned in the dossier_id field of POST /v1/docs/generate
      responses:
        "200":
          description: Dossier JSON with the metadata needed to verify it
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DossierResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Dossier not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          description: The dossier index entry or its stored content could not be read or decoded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpstreamError"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: >-
        API-key mode, for self-hosted single-operator deployments: the value
        of OPENCOMPLAI_API_KEY. A shared key authenticates the deployment as a
        whole, as the tenant named by OPENCOMPLAI_API_KEY_TENANT_ID (default
        oss-default). Ignored when OIDC_JWKS_URI is set. See
        docs/src/deployment/authentication.md.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        OIDC mode, for multi-user deployments (OIDC_JWKS_URI set): an RS256
        JWT verified against the identity provider's JWKS, with iss, aud, exp
        and nbf enforced. The token must also carry the tenant claim named by
        OIDC_TENANT_CLAIM. See docs/src/deployment/authentication.md.
...
  responses:
    Unauthorized:
      description: >-
        Credentials missing or invalid (POLICY_DENIED). The message says
        which: "Invalid or missing API key." in API-key mode; "Missing Bearer
        token.", "Invalid or expired JWT." or "Token missing tenant claim." in
        OIDC mode.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: >-
        Too many requests. After 10 failed authentication attempts from one
        client within 60 seconds (the defaults) further failures return 429
        RATE_LIMITED in the gateway envelope instead of 401. Separately, more
        than 300 requests of any kind from one client within 60 seconds (the
        default; the Docker Compose stack raises it to 1000) return the rate
        limiter's own body, {statusCode, error, message}, with a Retry-After
        header. Both limits are configurable.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/ErrorEnvelope"
              - $ref: "#/components/schemas/RequestRateLimitError"
...
    DependencyUnavailable:
      description: An upstream service could not be reached
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DependencyUnavailableError"
...
  schemas:
...
    UpstreamError:
      type: object
      required: [detail]
      description: >-
        Error body from an internal service, forwarded unchanged by the
        gateway. The services are FastAPI apps, so the body is
        {"detail": ...}; detail is a string, an object (for example
        {error_code, message, category, retryable}) or, for request validation
        failures, a list of error objects.
      properties:
        detail: {}
...
    AppendEventRequest:
      type: object
      required: [event_type, payload]
      properties:
        event_type:
          type: string
        payload:
          type: object
        signer_id:
          type: string
    AppendEventResponse:
      type: object
      required: [event_id, payload_hash, prev_hash]
      properties:
        event_id:
          type: string
        payload_hash:
          type: string
        prev_hash:
          type: string
...
    LedgerRootResponse:
      type: object
      required: [ledger_root_hash]
      properties:
        ledger_root_hash:
          type: string
          description: sha256:<hex> tip of the chain; the hash of the empty string when the ledger is empty
...
    GenerateDocsRequest:
      type: object
      required: [system_id, commit_ref]
      description: >-
        Body of the doc generator's POST /v1/docs/generate. The gateway has no
        schema of its own for this route and forwards the body as received.
      properties:
        system_id:
          type: string
        commit_ref:
          type: string
          description: Git commit reference; required, there is no default
        intended_purpose:
          type: string
          default: Not specified
        provider_name:
          type: string
          default: Unknown Provider
        compliance_target:
          type: string
          default: EU_AI_ACT
          description: >-
            Framework target recorded in the dossier. Must be EU_AI_ACT or
            NIST_AI_RMF; any other value fails with 500 SYSTEM_ERROR
        high_risk_presumption:
          type: boolean
          default: false
        allow_incomplete:
          type: boolean
          default: false
          description: >-
            Accept a dossier that fails the Annex IV schema check (200 with
            schema_valid false) instead of refusing it with 422
        training_data_description:
          type: [string, "null"]
        model_architecture:
          type: [string, "null"]
        performance_metrics:
          type: object
          additionalProperties:
            type: number
        known_limitations:
          type: array
          items:
            type: string
        human_oversight_measures:
          type: array
          items:
            type: string
        human_oversight:
          type: object
          additionalProperties: true
          description: >-
            Structured human-oversight declaration (roles, escalation,
            evidence references). Optional; forwarded untouched and validated
            by the downstream service.
        agent_inventory:
          type: object
          additionalProperties: true
          description: >-
            Declared agent inventory (agents, tools, mandate, delegation).
            Optional; forwarded untouched and validated by the downstream
            service.
        monitoring_approach:
          type: [string, "null"]
        incident_response_procedure:
          type: [string, "null"]
        metrics_appropriateness_rationale:
          type: [string, "null"]
        lifecycle_changes:
          type: array
          items:
            type: string
        change_log_reference:
          type: [string, "null"]
        harmonised_standards:
          type: array
          items:
            type: string
        alternative_solutions:
          type: [string, "null"]
        eu_declaration_of_conformity_ref:
          type: [string, "null"]
        post_market_monitoring_plan_ref:
          type: [string, "null"]
        post_market_monitoring_summary:
          type: [string, "null"]
        eval_report:
          type: [object, "null"]
          additionalProperties: true
          description: Full eval report as produced by the CLI; validated by the doc generator
        corroboration_report:
          type: [object, "null"]
          additionalProperties: true
          description: Full scan corroboration report; validated by the doc generator
    GenerateDocsResponse:
      type: object
      required: [dossier_id, bundle_checksum, status, duration_ms, schema_valid]
      properties:
        dossier_id:
          type: string
          description: Use it with GET /v1/docs/{dossier_id}
        bundle_checksum:
          type: string
        status:
          type: string
          enum: [generated]
          description: Always generated; a failure is reported as an error status instead
        duration_ms:
          type: integer
        signature:
          type: [string, "null"]
        schema_valid:
          type: boolean
          description: false only when allow_incomplete was set; otherwise an invalid dossier is a 422
        content_hash:
          type: [string, "null"]
          description: Content hash of the stored dossier
        ledger_event_id:
          type: [string, "null"]
          description: Ledger event that records the dossier
...